+++ shell.nix
@@ -7,15 +7,19 @@
 }@args:
 
 let
-  inherit (pkgs) cargo-deny openssl;
+  inherit (pkgs) cargo-deny openssl sqlite;
   shell = pimalaya.mkShell (removeAttrs args [ "pimalaya" ]);
 
 in
 shell.overrideAttrs (prev: {
-  LD_LIBRARY_PATH = pkgs.lib.makeLibraryPath [ openssl ];
+  LD_LIBRARY_PATH = pkgs.lib.makeLibraryPath [
+    openssl
+    sqlite
+  ];
 
   buildInputs = (prev.buildInputs or [ ]) ++ [
     cargo-deny
     openssl
+    sqlite
   ];
 })